+++ src/parsers/header.rs
@@ -11,25 +11,19 @@
 
 use std::collections::hash_map::Entry;
 
-use crate::{HeaderName, HeaderValue, Headers};
+use crate::{HeaderName, HeaderOffset, HeaderValue, Headers};
 
 use super::{
     fields::{
-        address::parse_address,
-        content_type::parse_content_type,
-        date::parse_date,
-        id::parse_id,
-        list::parse_comma_separared,
-        raw::{parse_and_ignore, parse_raw},
-        unstructured::parse_unstructured,
+        address::parse_address, content_type::parse_content_type, date::parse_date, id::parse_id,
+        list::parse_comma_separared, raw::parse_raw, unstructured::parse_unstructured,
     },
     message::MessageStream,
 };
 
 #[derive(Debug, PartialEq)]
 pub enum HeaderParserResult<'x> {
-    Supported(HeaderName),
-    Unsupported(&'x [u8]),
+    Header(HeaderName<'x>),
     Lf,
     Eof,
 }
@@ -40,19 +34,26 @@ pub fn parse_headers<'x>(headers: &mut Headers<'x>, stream: &mut MessageStream<'
         stream.pos += bytes_read;
 
         match result {
-            HeaderParserResult::Supported(name) => {
-                let (is_many, parser) = HDR_PARSER[name as usize];
+            HeaderParserResult::Header(name) => {
+                let (is_many, parser) = HDR_PARSER[name.id() as usize];
 
+                let from_offset = stream.pos;
                 let value = parser(stream);
+                headers.offsets.push(HeaderOffset {
+                    name: name.clone(),
+                    start: from_offset,
+                    end: stream.pos,
+                });
+
                 if !value.is_empty() {
                     if is_many {
-                        match headers.entry(name) {
+                        match headers.values.entry(name.clone()) {
                             Entry::Occupied(mut e) => {
                                 if let HeaderValue::Collection(col) = e.get_mut() {
                                     col.push(value);
                                 } else {
                                     let old_value = e.remove();
-                                    headers.insert(
+                                    headers.values.insert(
                                         name,
                                         HeaderValue::Collection(vec![old_value, value]),
                                     );
@@ -63,11 +64,10 @@ pub fn parse_headers<'x>(headers: &mut Headers<'x>, stream: &mut MessageStream<'
                             }
                         }
                     } else {
-                        headers.insert(name, value);
+                        headers.values.insert(name, value);
                     }
                 }
             }
-            HeaderParserResult::Unsupported(_name) => parse_and_ignore(stream),
             HeaderParserResult::Lf => return true,
             HeaderParserResult::Eof => return false,
         }
@@ -100,12 +100,17 @@ pub fn parse_header_name(data: &[u8]) -> (usize, HeaderParserResult) {
                             if field.eq_ignore_ascii_case(HDR_NAMES[token_hash]) {
                                 return (
                                     bytes_read,
-                                    HeaderParserResult::Supported(HDR_MAP[token_hash]),
+                                    HeaderParserResult::Header(HDR_MAP[token_hash].clone()),
                                 );
                             }
                         }
                     }
-                    return (bytes_read, HeaderParserResult::Unsupported(field));
+                    return (
+                        bytes_read,
+                        HeaderParserResult::Header(HeaderName::Other(String::from_utf8_lossy(
+                            field,
+                        ))),
+                    );
                 }
             }
             b'\n' => {
@@ -133,24 +138,51 @@ pub fn parse_header_name(data: &[u8]) -> (usize, HeaderParserResult) {
     (bytes_read, HeaderParserResult::Eof)
 }
 
-/*fn set_unsupported(&mut self, stream: &mut MessageStream<'x>, name: &'x [u8]) {
-    if let Ok(name) = std::str::from_utf8(name) {
-        let value = parse_unstructured(stream);
-        if !value.is_empty() {
-            match self.others.entry(name) {
-                Entry::Occupied(mut e) => {
-                    if let HeaderValue::Collection(col) = e.get_mut() {
-                        col.push(value);
-                    } else {
-                        let old_value = e.remove();
-                        self.others.insert(name,HeaderValue::Collection(vec![old_value, value]));
-                    }
-                },
-                Entry::Vacant(e) => {e.insert(value);},
-            }
+impl<'x> HeaderName<'x> {
+    fn id(&self) -> u8 {
+        match self {
+            HeaderName::Subject => 0,
+            HeaderName::From => 1,
+            HeaderName::To => 2,
+            HeaderName::Cc => 3,
+            HeaderName::Date => 4,
+            HeaderName::Bcc => 5,
+            HeaderName::ReplyTo => 6,
+            HeaderName::Sender => 7,
+            HeaderName::Comments => 8,
+            HeaderName::InReplyTo => 9,
+            HeaderName::Keywords => 10,
+            HeaderName::Received => 11,
+            HeaderName::MessageId => 12,
+            HeaderName::References => 13,
+            HeaderName::ReturnPath => 14,
+            HeaderName::MimeVersion => 15,
+            HeaderName::ContentDescription => 16,
+            HeaderName::ContentId => 17,
+            HeaderName::ContentLanguage => 18,
+            HeaderName::ContentLocation => 19,
+            HeaderName::ContentTransferEncoding => 20,
+            HeaderName::ContentType => 21,
+            HeaderName::ContentDisposition => 22,
+            HeaderName::ResentTo => 23,
+            HeaderName::ResentFrom => 24,
+            HeaderName::ResentBcc => 25,
+            HeaderName::ResentCc => 26,
+            HeaderName::ResentSender => 27,
+            HeaderName::ResentDate => 28,
+            HeaderName::ResentMessageId => 29,
+            HeaderName::ListArchive => 30,
+            HeaderName::ListHelp => 31,
+            HeaderName::ListId => 32,
+            HeaderName::ListOwner => 33,
+            HeaderName::ListPost => 34,
+            HeaderName::ListSubscribe => 35,
+            HeaderName::ListUnsubscribe => 36,
+            HeaderName::Other(_) => 37,
+            HeaderName::Invalid => 38,
         }
     }
-}*/
+}
 
 #[cfg(test)]
 mod tests {
@@ -161,27 +193,30 @@ mod tests {
     #[test]
     fn header_name_parse() {
         let inputs = [
-            ("From: ", HeaderParserResult::Supported(HeaderName::From)),
+            ("From: ", HeaderParserResult::Header(HeaderName::From)),
             (
                 "receiVED: ",
-                HeaderParserResult::Supported(HeaderName::Received),
+                HeaderParserResult::Header(HeaderName::Received),
             ),
             (
                 " subject   : ",
-                HeaderParserResult::Supported(HeaderName::Subject),
+                HeaderParserResult::Header(HeaderName::Subject),
             ),
             (
                 "X-Custom-Field : ",
-                HeaderParserResult::Unsupported(b"X-Custom-Field"),
+                HeaderParserResult::Header(HeaderName::Other("X-Custom-Field".into())),
             ),
-            (" T : ", HeaderParserResult::Unsupported(b"T")),
+            (
+                " T : ",
+                HeaderParserResult::Header(HeaderName::Other("T".into())),
+            ),
             (
                 "mal formed: ",
-                HeaderParserResult::Unsupported(b"mal formed"),
+                HeaderParserResult::Header(HeaderName::Other("mal formed".into())),
             ),
             (
                 "MIME-version : ",
-                HeaderParserResult::Supported(HeaderName::MimeVersion),
+                HeaderParserResult::Header(HeaderName::MimeVersion),
             ),
         ];
 
@@ -235,7 +270,7 @@ static HDR_PARSER: &[(
     (false, parse_address),         // ListPost = 34,
     (false, parse_address),         // ListSubscribe = 35,
     (false, parse_address),         // ListUnsubscribe = 36,
-    (false, parse_raw),             // Unsupported = 37,
+    (true, parse_raw),              // Other = 37,
 ];
 
 static HDR_HASH: &[u8] = &[
@@ -254,18 +289,18 @@ static HDR_HASH: &[u8] = &[
 
 static HDR_MAP: &[HeaderName] = &[
     HeaderName::Date,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::Sender,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::Received,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::References,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::Cc,
     HeaderName::Comments,
     HeaderName::ResentCc,
     HeaderName::ContentId,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::ResentMessageId,
     HeaderName::ReplyTo,
     HeaderName::ResentTo,
@@ -273,54 +308,54 @@ static HDR_MAP: &[HeaderName] = &[
     HeaderName::ContentLanguage,
     HeaderName::Subject,
     HeaderName::ResentSender,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
     HeaderName::ResentDate,
     HeaderName::To,
     HeaderName::Bcc,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::ContentTransferEncoding,
     HeaderName::ReturnPath,
     HeaderName::ListId,
     HeaderName::Keywords,
     HeaderName::ContentDescription,
     HeaderName::ListOwner,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::ContentType,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::ListHelp,
     HeaderName::MessageId,
     HeaderName::ContentLocation,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
     HeaderName::ListSubscribe,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
     HeaderName::ListPost,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::ResentFrom,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
     HeaderName::ContentDisposition,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::InReplyTo,
     HeaderName::ListArchive,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::From,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
     HeaderName::ListUnsubscribe,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
-    HeaderName::Unsupported,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
+    HeaderName::Invalid,
     HeaderName::MimeVersion,
 ];